requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
bcrypt==4.0.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0